+++ src/go/rpk/pkg/cli/cmd/redpanda/admin/brokers/brokers.go
@@ -12,12 +12,6 @@
 package brokers
 
 import (
-	"fmt"
-	"strconv"
-
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/api/admin"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/out"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
@@ -35,129 +29,4 @@ func NewCommand(fs afero.Fs) *cobra.Command {
 		newRecommissionBroker(fs),
 	)
 	return cmd
-}
-
-func newListCommand(fs afero.Fs) *cobra.Command {
-	return &cobra.Command{
-		Use:     "list",
-		Aliases: []string{"ls"},
-		Short:   "List the brokers in your cluster",
-		Args:    cobra.ExactArgs(0),
-		Run: func(cmd *cobra.Command, _ []string) {
-			p := config.ParamsFromCommand(cmd)
-			cfg, err := p.Load(fs)
-			out.MaybeDie(err, "unable to load config: %v", err)
-
-			cl, err := admin.NewClient(fs, cfg)
-			out.MaybeDie(err, "unable to initialize admin client: %v", err)
-
-			bs, err := cl.Brokers(cmd.Context())
-			out.MaybeDie(err, "unable to request brokers: %v", err)
-
-			headers := []string{"Node-ID", "Num-Cores", "Membership-Status"}
-
-			args := func(b *admin.Broker) []interface{} {
-				ret := []interface{}{b.NodeID, b.NumCores, b.MembershipStatus}
-				return ret
-			}
-			for _, b := range bs {
-				if b.IsAlive != nil {
-					headers = append(headers, "Is-Alive", "Broker-Version")
-					orig := args
-					args = func(b *admin.Broker) []interface{} {
-						return append(orig(b), *b.IsAlive, b.Version)
-					}
-					break
-				}
-			}
-			tw := out.NewTable(headers...)
-			defer tw.Flush()
-			for _, b := range bs {
-				tw.Print(args(&b)...)
-			}
-		},
-	}
-}
-
-func newDecommissionBroker(fs afero.Fs) *cobra.Command {
-	return &cobra.Command{
-		Use:   "decommission [BROKER ID]",
-		Short: "Decommission the given broker",
-		Long: `Decommission the given broker.
-
-Decommissioning a broker removes it from the cluster.
-
-A decommission request is sent to every broker in the cluster, only the cluster
-leader handles the request.
-`,
-		Args: cobra.ExactArgs(1),
-		Run: func(cmd *cobra.Command, args []string) {
-			broker, err := strconv.Atoi(args[0])
-			out.MaybeDie(err, "invalid broker %s: %v", args[0], err)
-			if broker < 0 {
-				out.Die("invalid negative broker id %v", broker)
-			}
-
-			p := config.ParamsFromCommand(cmd)
-			cfg, err := p.Load(fs)
-			out.MaybeDie(err, "unable to load config: %v", err)
-
-			cl, err := admin.NewClient(fs, cfg)
-			out.MaybeDie(err, "unable to initialize admin client: %v", err)
-
-			b, err := cl.Broker(cmd.Context(), broker)
-			out.MaybeDie(err, "unable to initialize admin client: %v", err)
-
-			// Old brokers (< v22.1) don't have maintenance mode, so we must
-			// check if b.Maintenance is not nil.
-			if b.Maintenance != nil && b.Maintenance.Draining {
-				out.Die(`Node cannot be decommissioned while it is in maintenance mode.
-Take the node out of maintenance mode first by running: 
-    rpk cluster maintenance disable %v`, broker)
-			}
-
-			err = cl.DecommissionBroker(cmd.Context(), broker)
-			out.MaybeDie(err, "unable to decommission broker: %v", err)
-
-			fmt.Printf("Success, broker %d has been decommissioned!\n", broker)
-		},
-	}
-}
-
-func newRecommissionBroker(fs afero.Fs) *cobra.Command {
-	return &cobra.Command{
-		Use:   "recommission [BROKER ID]",
-		Short: "Recommission the given broker if it is still decommissioning",
-		Long: `Recommission the given broker if is is still decommissioning.
-
-Recommissioning can stop an active decommission.
-
-Once a broker is decommissioned, it cannot be recommissioned through this
-command.
-
-A recommission request is sent to every broker in the cluster, only
-the cluster leader handles the request.
-
-`,
-		Args: cobra.ExactArgs(1),
-		Run: func(cmd *cobra.Command, args []string) {
-			broker, err := strconv.Atoi(args[0])
-			out.MaybeDie(err, "invalid broker %s: %v", args[0], err)
-			if broker < 0 {
-				out.Die("invalid negative broker id %v", broker)
-			}
-
-			p := config.ParamsFromCommand(cmd)
-			cfg, err := p.Load(fs)
-			out.MaybeDie(err, "unable to load config: %v", err)
-
-			cl, err := admin.NewClient(fs, cfg)
-			out.MaybeDie(err, "unable to initialize admin client: %v", err)
-
-			err = cl.RecommissionBroker(cmd.Context(), broker)
-			out.MaybeDie(err, "unable to recommission broker: %v", err)
-
-			fmt.Printf("Success, broker %d has been recommissioned!\n", broker)
-		},
-	}
 }
